added error handling and messagebox

+++ main.py
@@ -1,5 +1,6 @@
 from PyQt5 import QtWidgets, uic, QtGui, QtCore
-from PyQt5.QtWidgets import QWidget
+from PyQt5.QtGui import QIcon
+from PyQt5.QtWidgets import QWidget, QMessageBox, qApp
 import sys
 import os
 
@@ -8,10 +9,12 @@ class App(QWidget):
     def __init__(self):
         super().__init__()
         self.path = 'Notes'
-        self.NOTE = 'Note №'
+        self.NOTE = 'Note '
+        self.number_of_note = 1
         self.ui = None
         self.notes_list = None
         self.model = QtGui.QStandardItemModel(self)
+        qApp.setStyleSheet("QMessageBox{background-color:rgb(255, 245, 124);}")
         self.start()
 
     def start(self):
@@ -23,7 +26,7 @@ class App(QWidget):
     # set up ListView widget
     def setuplistveiw(self):
         self.model.clear()
-        self.notes_list = self.ListOfNotes()
+        self.notes_list = os.listdir(self.path)
         for note in self.notes_list:
             item = QtGui.QStandardItem(note)
             item.setData(note)
@@ -33,57 +36,68 @@ class App(QWidget):
 
     # set up buttons
     def set(self):
-        self.ui.btn1.clicked.connect(lambda: self.NewNote(number_of_note=1))
+        self.ui.btn1.clicked.connect(lambda: self.NewNote())
         self.ui.btn2.clicked.connect(lambda: self.EditNote())
         self.ui.btn3.clicked.connect(lambda: self.DeleteNote())
         self.ui.listView.clicked.connect(lambda: self.item_clicked())
+        # self.ui.listView.mouseDoubleClickEvent(lambda: self.rename_note())
+
+    def create_messagebox(self, text):
+        messagebox = QMessageBox()
+        messagebox.setText(text)
+        messagebox.setWindowTitle('Error')
+        messagebox.setWindowIcon(QIcon('icons/warning.png'))
+        messagebox.setStandardButtons(QMessageBox.Ok)
+        messagebox.exec_()
 
     # get name of selected note
     @QtCore.pyqtSlot()
     def item_clicked(self):
         index = self.ui.listView.selectionModel().currentIndex()
         file_name = index.data()
-        print(file_name)
         with open(f"Notes/{file_name}", 'r', encoding='utf-8') as file:
             self.ui.textEdit.setPlainText(file.read())
         return file_name
 
     # creating a new note
-    def NewNote(self, number_of_note=1):
-        file_name = f'{self.NOTE} {number_of_note}.txt'
-        text = self.ui.textEdit.toPlainText()
-        if file_name in os.listdir(self.path):
-            file_name = f"{self.NOTE} {number_of_note + len(os.listdir(self.path))}.txt"
-        with open(f'Notes/{file_name}', 'w', encoding='utf-8') as file:
-            file.write(text)
+    def NewNote(self):
+        if self.ui.textEdit.toPlainText() != '':
+            file_name = f'{self.NOTE} {self.number_of_note}.txt'
+            if file_name in os.listdir(self.path):
+                file_name = f"{self.NOTE} {self.number_of_note + len(os.listdir(self.path))}.txt"
+                print(file_name)
+            text = self.ui.textEdit.toPlainText()
+            with open(f'Notes/{file_name}', 'w', encoding='utf-8') as file:
+                file.write(text)
 
-        # update list of notes in ListView widget
-        item = QtGui.QStandardItem(file_name)
-        item.setData(file_name)
-        self.model.appendRow(item)
+            # update list of notes in ListView widget
+            item = QtGui.QStandardItem(file_name)
+            item.setData(file_name)
+            self.model.appendRow(item)
+            self.ui.textEdit.clear()
+        else:
+            self.create_messagebox(text='Empty note')
 
-    # editing an existing note
+    # edit an existing note
     def EditNote(self):
-        list_of_notes = self.ListOfNotes()
-        print(list_of_notes)
-        # note_for_editing = int(input('Enter number of note: ')) - 1
-        # with open(f"Notes/{list_of_notes[note_for_editing]}", encoding='utf-8') as file:
-        #     text = file.read()
-        #     print("Content of note: \n", text)
-        #     new_text = input(f"New content: ")
-        #     with open(f"Notes/{list_of_notes[note_for_editing]}", 'w', encoding='utf-8') as newfile:
-        #         newfile.write(new_text)
-
-    # get an existing notes list
-    def ListOfNotes(self):
-        list_of_notes = os.listdir(self.path)
-        return list_of_notes
+        try:
+            new_text = self.ui.textEdit.toPlainText()
+            with open(f'Notes/{self.item_clicked()}', 'w', encoding='utf-8') as file:
+                file.write(new_text)
+            self.setuplistveiw()
+            self.ui.textEdit.clear()
+        except:
+            self.create_messagebox(text='Select a note')
 
     # delete existing note
     def DeleteNote(self):
-        choice = self.item_clicked()
-        os.remove(f'{self.path}/{choice}')
-        self.setuplistveiw()
+        try:
+            choice = self.item_clicked()
+            os.remove(f'{self.path}/{choice}')
+            self.ui.textEdit.clear()
+            self.setuplistveiw()
+        except:
+            self.create_messagebox(text='Select a note')
 
 
 if __name__ == '__main__':
